+++ apps/health_campaign_field_worker_app/lib/utils/utils.dart
@@ -444,21 +444,6 @@ int getSyncCount(List<OpLog> oplogs) {
   return count;
 }
 
-
-class TotalDetails {
-  num noOfHousesVisited;
-  num noOfHousesSprayed;
-  num noOfHousesNotSprayed;
-  num bottlesUsed;
-  num noOfRemainingBottles;
-  TotalDetails({
-    this.noOfHousesVisited = 0,
-    this.noOfHousesSprayed = 0,
-    this.noOfHousesNotSprayed = 0,
-    this.bottlesUsed = 0,
-    this.noOfRemainingBottles = 0,
-  });
-
 class LocalizationParams {
   static final LocalizationParams _singleton = LocalizationParams._internal();