+++ packages/digit_data_model/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.3
+
+* Added localization table
+* Updated individual local search results with user uuid mapping
+
 ## 1.0.3-dev.4
 
 * Updated individual local search results with user uuid mapping